Update view to delete and update a todo

+++ todos/views.py
@@ -4,6 +4,7 @@ from rest_framework.response import Response
 from rest_framework import status
 from .models import Todo
 from .serializers import TodoSerializer
+from django.contrib.auth.models import User
 
 @api_view(['DELETE', 'PUT'])
 def delete_update_todo(request, pk):
@@ -14,15 +15,23 @@ def delete_update_todo(request, pk):
 
     # delete a single todo
     if request.method == 'DELETE':
-        todo.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if todo.user == request.user :
+            todo.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response(status=status.HTTP_403_FORBIDDEN)
+    
     #  update of a single todo
     elif request.method == 'PUT':
-        serializer = TodoSerializer(todo, data = request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status= status.HTTP_204_NO_CONTENT)
+        if todo.user == request.user:
+            serializer = TodoSerializer(todo, data = request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status= status.HTTP_204_NO_CONTENT)
+        else:
+            return Response(status=status.HTTP_403_FORBIDDEN)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
 
 @api_view(['GET', 'POST'])
 def get_post_todo(request):
@@ -31,6 +40,7 @@ def get_post_todo(request):
         todos = Todo.objects.all()
         serializer = TodoSerializer(todos, many=True)
         return Response(serializer.data)
+    
     # insert a new record for a todo
     elif request.method == 'POST':
         data = {
@@ -38,7 +48,7 @@ def get_post_todo(request):
         }
         serializer = TodoSerializer(data=data)
         if serializer.is_valid():
-            serializer.save()
+            serializer.save(user=request.user)
             return Response(serializer.data, status = status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
